added: Riot.js 4 keyed and not keyed examples
