--- frameworks/non-keyed/riot/src/app.html
@@ -1,124 +1,0 @@
-/*global riot */
-<app>
-  <div class='container'>
-    <div class='jumbotron'>
-      <div class='row'>
-        <div class='col-md-6'>
-          <h1>riot</h1>
-        </div>
-        <div class='col-md-6'>
-          <div class='row'>
-            <div class='col-sm-6 smallpad'>
-              <button type='button' class='btn btn-primary btn-block' id='run' onclick={ methods.run }>
-                Create 1,000 rows
-              </button>
-            </div>
-            <div class='col-sm-6 smallpad'>
-              <button type='button' class='btn btn-primary btn-block' id='runlots' onclick={ methods.runLots }>Create 10,000 rows
-              </button>
-            </div>
-            <div class='col-sm-6 smallpad'>
-              <button type='button' class='btn btn-primary btn-block' id='add' onclick={ methods.add }>
-                Append 1,000 rows
-              </button>
-            </div>
-            <div class='col-sm-6 smallpad'>
-              <button type='button' class='btn btn-primary btn-block' id='update' onclick={ methods.update }>Update every 10th row
-              </button>
-            </div>
-            <div class='col-sm-6 smallpad'>
-              <button type='button' class='btn btn-primary btn-block' id='clear' onclick={ methods.clear }>
-                Clear
-              </button>
-            </div>
-            <div class='col-sm-6 smallpad'>
-              <button type='button' class='btn btn-primary btn-block' id="swaprows" onclick={ methods.swapRows } >
-                Swap Rows
-              </button>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-    <table class='table table-hover table-striped test-data'>
-      <tbody>
-        <tr each={ rows } class={ danger: id == selected } no-reorder>
-          <td class='col-md-1'>{ id }</td>
-          <td class='col-md-4'>
-            <a onclick={ methods.select }>{ label }</a>
-          </td>
-          <td class='col-md-1'>
-            <a onclick={ methods.remove }><span class='glyphicon glyphicon-remove' aria-hidden='true'></span></a>
-          </td>
-          <td class='col-md-6'></td>
-        </tr>
-      </tbody>
-    </table>
-    <span class='preloadicon glyphicon glyphicon-remove' aria-hidden='true'></span>
-  </div>
-<script>
-  const store = opts.store
-
-  this.methods = {
-    add: () => {
-      startMeasure('add')
-      store.add()
-      this.rows = store.data
-      this.selected = store.selected
-      stopMeasure()
-    },
-    remove: (e) => {
-      startMeasure('remove')
-      store.delete(e.item.id)
-      this.rows = store.data
-      this.selected = store.selected
-      stopMeasure()
-    },
-    select: (e) => {
-      startMeasure('select')
-      store.select(e.item.id)
-      this.rows = store.data
-      this.selected = store.selected
-      stopMeasure()
-    },
-    run: () => {
-      startMeasure('run')
-      store.run()
-      this.rows = store.data
-      this.selected = store.selected
-      stopMeasure()
-    },
-    update: () => {
-      startMeasure('update')
-      store.update()
-      this.rows = store.data
-      this.selected = store.selected
-      stopMeasure()
-    },
-    runLots: () => {
-      startMeasure('runLots')
-      store.runLots()
-      this.rows = store.data
-      this.selected = store.selected
-      stopMeasure()
-    },
-    clear: () => {
-      startMeasure('clear')
-      store.clear()
-      this.rows = store.data
-      this.selected = store.selected
-      stopMeasure()
-    },
-    swapRows: () => {
-      startMeasure('swapRows')
-      store.swapRows()
-      this.rows = store.data
-      this.selected = store.selected
-      stopMeasure()
-    }
-  }
-
-  this.rows = store.data
-  this.selected = store.selected
-</script>
-</app>
