--- frameworks/non-keyed/riot/src/main.es6.js
@@ -1,7 +1,0 @@
-import riot from 'riot'
-import Store from './store'
-import './app.html'
-
-riot.mount('app', {
-  store: new Store()
-});
